Remove references to the MIPS syscall instruction from the cross-platform recompiler driver, cpu_asm.c.
Overhaul the MIPS stub and emitter so that they can be compiled with or without load delays, as well as with MIPS32 r1 or r2 instructions.

Remove references to the serial port in #define SERIAL_TRACE*, since the serial port may or may not be the trace output of choice for a platform.

Remove some unused functions and macros.